feat: use Closure Compiler's types.
This code uses Closure Compiler's JSType & friends to generate a
corresponding `.d.ts` file. This allows for much more precision as to how
Closure perceives the type system of a JS program. It also simplifies the
code generator by walking a structure that is mostly already suitable for the
task (using types instead of syntactical structures that need to be converted
into a reasonable program).

This also introduces golden file based tests, and removes code from the previous
approach.