feat: mismatched inherited static properties.
In TypeScript, the "static" side of classes must correctly inherit from
any of its parents. In Closure, the methods are completely separate and
do not have to match. To paper over the difference, emit overloads so
that TypeScript allows the inheritance.

This pretends functions exist that do not, and it will fail for
functions that only differ in return type, but it should be good enough
for most code, for now.

Fixes #120.