fix: inner classes, enums and interfaces are emitted.
Previously, we expected an explicit goog.provide for inner classes,
enums and interfaces, but according to style guide such requirers are
forbidden.

Now we do a second pass over the properties of default exported classes
and emit the inner properties that are not static methods or fields into
a namespace.

Finally, while against style guide, in some cases internal classes are
explicitly goog.provided, so we add check to make sure they are no
repeated exports.