feat: prevent structural typing w/ private fields.
When a type has private fields, add a special field called
`private noStructuralTyping_: any;` to avoid accidentally applying structural
typing to a type that does not support that.

This only works for `class`es - non-structural `interface`s in Closure will
silently allow structural typing for the time being, because we cannot add
fields to them. I guess we could warn about this in the future on the tsc side.

Also move all tests for `@private` annotations into their own file, so later
maintainers have a chance to find that code.