fix: always generate absolute names.
This makes it much less likely that a type or package gets shadowed by another
type or package with the same name somewhere in the namespace hierarchy.

There is a remaining corner case around re-declared global names like `Element`
or `Object`. The fix would be to alias *all* global names, but that risks
aliasing global names from Closure that do not exist in TypeScript.

Fixes #144.