emit $jscomp.scope objects.
When using the pattern:

goog.provide('foo.Bar');
goog.scope(function() {
  /** @constructor */
  var A = function() {};
  /** @const */ foo.Bar = A;
}

foo.Bar is a valid type symbol, (i.e. it can be used as
/** @type {foo.Bar} */). However that is just a type alias for the
synthetic $jscomp.scope.A type symbol. Moreover, the dealiasing happens
in InlineAliases pass, and the aliasing map is not accessible for
further use.

Previously, we tried to revert the aliasing for previous goog.module,
but that did not handle general goog.scope. Now, we just emit all
$jscomp.scope type symbols used throughout the inputs.

Fixes a previous bug in aggregating typesUsed, where it did not account
for union types (null|Foo).