Emit typedefs by name instead of unrolling them.
Summary:
When walking types, Closure internally inlines the type aliases. Before
this change we did the naive thing, which meant that the emitted type
was expanded.

For example for closure:
  /** @typedef {{foo: number}} */ T;
  /** @const {T} */ var x;

We would have emitted:
  typedef T = {foo: number};
  var x: {foo: number};

This strategy does not work in presence of recursive typedefs. The
commit changes the output to named typedefs for the ones we know for
sure we have emitted so far.

The example above will now look like:
  typedef T = {foo: number};
  var x: T;

Closes #286

Reviewers: evmar

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D147