Initial commit of a new command line tool gents.
Summary:
Gents is a command line tool used to transpile Closure annotated
ES6/ES5 JavaScript to TypeScript. It uses the declared type information
provided in the JSDoc to determine the Typescript types.

Currently, we reuse JsdocToEs6TypedConverter as a CompilerPass to type
annotate the source code. We also reuse the code generation for ES6_Typed
provided by the Closure compiler as a first attempt to generate TypeScript

Reviewers: mprobst, rkirov

Reviewed By: mprobst, rkirov

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D150