Implement class constructor conversion from ES5 to ES6.
Summary:
This converts class constructors declared in ES5 to the new
class syntax introduced in ES6. We only convert explicitly
annotated constructors to not eagerly infer too much information.

Conversion is handled as a separate `CompilerPass` to leave
open the possibility of migrating the code into Closure Compiler.

This is still a WIP and does not produce valid TypeScript yet.

Reviewers: rkirov, mprobst

Reviewed By: mprobst

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D167