fix NPE when visiting super classes.
Summary:
When the compilation unit has errors it can happen that the supertype is
UNKNOWN, with is a non-null type with a null constructor.

This case is not tested, because the prerequisite for obtaining super
type being UNKNOWN is to have an invalid program. This cannot be
constructed by our test setup.

Reviewers: mprobst

Reviewed By: mprobst

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D180