feat: (non) strictdeps.
Summary:
Add a test for the accidentally introduced strictdeps feature. In this
mode, Clutz generates module declarations only for the immediate roots,
but still generates type definitions for types reachable from the roots'
types.

Also adds an option to disable this mode and go back to the previous
behaviour of generating types and modules for all inputs, to ease
migration to the new mode.

Reviewers: rkirov, renez

Reviewed By: rkirov, renez

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D200