Transpile comments from JS to TS in gents.
Summary:
Resolves #310.

Documentation is preserved in the translation from JavaScript to TypeScript.

Comments are attached to the AST in "comment blocks". Comments directly
before code is attached directly to the code. Otherwise, they are considered
"floating" and will be attached to the root node as an EMPTY node.

Note that currently, this does not support endline comments well.

Reviewers: rkirov, mprobst

Reviewed By: mprobst

Subscribers: typescript-eng

Differential Revision: https://reviews.angular.io/D232