fix: Don't rename default export-ed values.
Summary:
Currently, if the name of a value does not match the module's suffix,
gents assigns the default export to a new variable that does. Remove
this behavior since it is unnecessary.

Assume the source is from `goog.module('package.A');`

The output would be:

Before:

```typescript

class Foo {}

export default const A = foo;
```

After:

```typescript

class Foo{}

export default Foo;
```