fix: Make newline insertion more consistent.
Summary:
Currently newline insertion breaks when the previous node ends with a
semi-colon, due to Closure Compiler's automatic semi-colon insertion.

That is, Gents outputs:
```typescript
const foo = 4
  ;
class Bar{}
```
which gets formatted by clang-format to:
```typescript
const foo = 4;
class Bar{}
```

This PR changes the insertion logic to insert a newline before the node
is written, so that it does not fall victim to semi-colon insertion.