Support enums nested in typedefs that are defined in externs and provided.
This surprising pattern is from an actual piece of code encountered in the wild.
It is an obscure corner case, and it is unclear whether the NoType check is
correct or might trigger in other programs. I will try this change out on a
larger code base and see if it breaks things.