Upgrade clutz tests to TS2.4. (#542)
Remove 'Promise' from the list of platform APIs for which we generate
types. This means that clutz will reuse the Promise definition from
lib.d.ts, instead of attempting to translate closure's definition.

Closure already uses TTE's for the types of Promise and we have to
resort to hard-coding the translation.

Also, fix a test that failed in the presence of properly covariant
promises.