Special case `entries()` methods.
This special cases methods with type (in TS notation) `entries():
IterableIterator<K|V>` to be emitted as `entries():
IterableIterator<K|V]>`. That is, instead of an iterator over a union of
key and value, it returns an iterator over a two-tuple of key and value.

This precisely represents the type of `entries()` methods in map-like
structures, which is a very common use case. Without the special case,
the union type breaks usage of the keys and values.