Add support for emitting extending/implementing an unresolved base.
Adds an empty ClutzMissingBase class to closure.lib.d.ts for the case
that closure fails to resolve the base type. I could not reproduce this
code path in test, but have seen it in the wild. Closure errs on simple
attempts to extend a non-resolved base symbol and in partial compilation
we can use the forward declared name.