Adds enum support in gents.
For string and number enums convert them to TypeScript, all other
types of enums are left as object literals.

The comment rewriting pass does not communicate with the pass that
modifies the AST, so it is not easy to decide whether to strip @enum
jsdocs or not. We can either always remove them or never remove them.

In this change, we never remove them, because there are still a number
of cases where we cannot rewrite to TS enum, so the comments keep
important information for the users.