Do not use typedefs that are not emitted.
When Clutz emits any type, it checks if there's a typedef for that
specific type, and emits the typedef instead of the underlying type.
That generally works, though it has somewhat odd consequences. For
example, a base library might end up using a typedef defined in an
entirely different, unrelated library.

This was caused by the fix for #286, in b66e265ed5. It might be worth
pursuing a better fix for the original issue of mutually recursive type
definitions.

Unrelated to that though, the current behaviour breaks in combination
with `skipEmitRegExp`. Clutz can end up using typedefs referencing code
whose typedefs are not emitted due to `skipEmitRegExp`, which then
reference non-existing code. This change mitigates the problem by never
picking up those typedefs in the first place.