Prefer earlier typedef names over latter ones.
Closure sources are sorted in dependency order during compilation. With
the current setup, due to `getAllSymbols` returning symbols in order of
their definition, typedefs that are found later in the sources always
win over typedefs that are found earlier.

This has no semantic impact, but it a bit odd, as this means base
libraries will prefer to use typedefs from the very leaf libraries.
This change flips the preference around by always preferring the earlier
typedef over a later one.