When parsing partial inputs, closure doesn't know the proper name of an imported symbol, so it emits it as part of the current module.
This PR walks the closure ast to grab the module id of the current file as well as the id of the import to construct a map of unresolved symbol names to resolved symbol names, and uses it to replace unresolved names when emitting NoResolvedType nodes.

To avoid reparsing the inputs, this PR introduces InitialParseRetainingCompiler, which clones and retains the ast of the inputs before any compiler passes have run.