hack around missing base class in partial mode for goog.deferred.Async
This class @extends Thenable, which means it should inherit a TTE type for
its .then() implementation.  But in partial mode, we don't see the base
class so we can't see that type.

We considered making it just infer to some broad any-based type, but in
in this particular case (the implementation of .then), if you lose the
TTE types then you have to manually annotate all the parameters to your
callback, so we deemed it ok to have a specific hack to see how common
this problem is.