Add a multifile partial clutz test.
Adds a test case for compiling 3 javascript files (1 importing from the other 2) individually, then running a DeclarationSyntaxTest on the result to ensure that Clutz produces correct output.

As a result, an error in ImportRenameMapBuilder was discovered where the imported symbols were being rename to modulename *underscore* symbolname, instead of the correct modulename *dot* symbolname.