Fix for goog.provides with inferred symbols.
Inferred symbols inside namespace provides are handled specially.

Previously, we had a fix for this issue, but only turned it on for
goog.modules. Now this is turned on for all inputs. Previous concerns
do not appear at least in the test suite.

This issue is more appearant with incremental clutz, because at least in
d3's case an ambient extern was masking the issue. That extern is no
longer visible in incremental scenario.

Note: this affects both modes (incremental and total) of clutz.