Add a partial_goog_base.js file that is used for all incremental clutz.
Knowing that 'goog.abstractMethod' is a function allows for incremental
clutz to emit 'C.prototype.fn = goog.abstractMethod' as a method and not
a field. TypeScript asserts that methods can only override other
methods, so this is crucial knowledge for incremental clutz.

To achieve this we add an extra partial version of base.js -
partial_goog_base.js that will be passed to every partial clutz
compilation. It will not affect the output in other ways than allowing
inference works to work better in its presence.

Also, make sure we do not emit optional varArgs as such are not legal
TypeScript.