Fix import renaming for es6 class syntax.
Using es6 syntax instead of /** @constructor */ annotations results in a different display name for a missing type in partial mode.  In an es6 the type is `foo` instead of the annoation method's `module$contents$foo`.  This PR modifies ImportRenameMapBuilder to produce mappings for both types of display name.

This PR also fixes an issue when importing an object, which contains the type to be used (ie imported foo, but used as `class quux extends foo.bar`.  In that case the displayname is of the form `foo.bar` or `module$contents$foo.bar`, but the importRenameMap only has a rename for the base of the display name.  To fix, the logic is modified to look for replacements of the portion of the display name up to the first period (if any).