mark partial overridden function parameters as optional
When a method @override's a superclass method, and we can't see the
superclass due to partial mode, we can't know whether the parameters
are optional or not.  Try to detect when this happens (when all the
parameters are type {?}) and defensively mark them all as optional.

This change is walking a fine line between multiple competing concerns.
For example, we don't want to rely on specifically @override because
of @inheritDoc, and we also want to do the right thing when the
superclass is visible.  See the test case for some examples.