Gents translates ES6 classes with @interface annotation to interfaces.
Some closure users prefer to write /** @interface */ class IFoo, instead
of the more common function syntax. The intent is nevertheless for this
to be a pure interface and not class, so gents will correspondingly
convert that.

Some complications arrise from adding @extends and ES6's class extend
keyword and method/field definitions.

Closes #630,627