use a sorted List for class template parameters
We pass the template parameters of a class down through a whole
train of functions, ultimately to grab the "first" parameter for
special-casing Promises.  Unfortunately we were using a HashSet for
this, which meant that we could get the second parameter instead
of the first depending on how the HashSet worked out.

Rather than just making this a LinkedHashSet, I though I might as
well make it a List because that is a more appropriate type.  In
doing so I found one place where we were accidentally passing the
list of class template parameters in to a function that was expecting
something different.  I made my best guess at the behavior we actually
wanted there.