use whitelists of known global types
When code refers to e.g. HTMLElement, we want to ensure that we
emit HTMLElement and not o_o.clutz.HTMLElement.  In incremental
mode the only thing we have to go on is the "HTMLElement" text
itself.

This change adds lists of:
1) all symbols declared by lib.d.ts
2) all symbols declared in externs that we *don't* want externs for

With those in hand, we can now assume that any symbol that isn't
declared refers to a global, and we can also avoid emitting externs
for symbols that TypeScript already declares.

This fixes one important test (which had the wrong reference to "Node")
and also removes some externs emit for some types that TypeScript
now covers.