add a whitelist for aliasing goog.provides
If a file A has
  goog.provide('mything');
  mything = 3;
and a file B has
  goog.provide('mything.subnamespace');
  mything.subnamespace = 4;

Then the code is nonsense but clutz has a workaround anyway.

In partial mode we look at A without knowledge of B, so we cannot apply
the workaround without providing a global set of broken inputs like
these.  This change adds support for the global set and hooks it up to
a command-line flag.