Support specialCaseMapEntries in incremental mode.
specialCaseMapEntries does some type inspection to make sure this is
indeed the special cased method. Those were write with the assumption
that externs are present, which is not true in incremental mode.

Same checks are now done using type names instead of resolved symbols.
There is no loss of safety.

Fixes #697