Incremental clutz should not add root depgraph externs to the compilation externs.
Depgraph roots can be exaggerated due to a mechanism of "exporting" for
js_library bazel rules. Both sources and externs should only be
filtering with the depgraph roots and not adding new targets. So far
this was true for sources and this change will make it true for externs.

Slight refactoring of tests. Switches to more descriptive names for depgraph files.