Fixes emit in partial mode on module local typedefs used in generics. (#702)
Usually closure inlines typedefs, but in partial mode with generics
involved it only resolves the name to $module$contents. Our secondary
types used pass has to deal with finding the typedef definition and
emitting it correctly.

We already had code that dealt with one-off module typedefs, so I
refactored that around so it can be reused.