Allow clutz to handle ES6 generator functions. (#717)
In closure generator functions implicitly are defined as functions that
return a Generator<T> type. In TypeScript, the closest thing to
Generator type is the IterableIterator.

The difference boils down to the the two methods return and throw, that
are part of the Generator interface in closure, but in TS they are
optional methods on the Iterator<T> type.

This change makes it so that we emit Genarator as IterableIterator in
the .d.ts. Tests added for both partial and total mode.

Special care for not emitting builtins Iterator and Generator, which
appear even if the externs they are defined in are not passed to the
compilation.