Make sure to use enum alias even if the original enum came from a genfile.
For some reason when this happened, we didn't get the correct qualified name, and this is a combination of having an enum aliased to another enum in a different file that was autogenerated.

The test case added is actually not enough to test this, but it was the way that many classes in the code base were fixed that should allow the tap train to pass.