Add support to read JS files inside ZIP files.
The use case for this is allowing the use of J2CL (see google/jsinterop-base).
J2CL bazel rules generate a ZIP file which will need declaration files generated, but clutz didn't support that.

There is documentation on how to generate ZIP files and golden test files in the *_in_zip.dts files
as well as the test helper function that gets the golden File instance.