Emit object literal types when visiting objects.
Previously, we just emitted GlobalObject (alias for Object). However, we
do have enough information to emit an objectl literal type, even though
 closure does not report those types as proper record types.

Especially for inferred namespaces this option is preferable to the
opaque object type.

For a real opaque 'Object' types instead of 'GlobalObject', clutz now
emits '{}' which is equivalent in TypeScript.