Do not consider simpler default obj lit exports as default exports.
Gents tries to not use default exports to account for an internal style
rule to not use default exports. However, some default exports are
trivial obj literals that just safe typing for the author.

class C {}
class D {}

exports = {C, D};

After this change gents will translate this file as if it was writen in
the equivalent style of:

exports.C = C;
exports.D = D;

Closes #527