Fixes exporting typedefs through obj lit.
Previously this caused an NPE due to missing nodes. It appears we held
back unrolling

```
let x = 0;
export {x};
```

into

```
export let x = 0;
```

because of possibility of mutable bindings. I think the vast majority of
bindings are immutable (for example typedefs), so this cl reverts this
decision. It prefers the simpler syntactic output for the majority,
while accepting that in some cases the translation is not semantically
 accurate.