reformat the README
This uses the google-internal markdown formatter as a one-time
reformat of the README file.  Because it is google-internal I
cannot hook it into the build, but I figure at least a one-time
formatting pass here will make the file a little less gross for
subsequent edits.

(I also investigated using prettier.io's formatter but it wants to
rewrite the JS snippets within the README too.)