Remove splitting classes into C (static side) and C_instance
This was needed to support @constructor es5 classes that had custom
semantics with respect to statics in closure. With this change some
closure code can produce .d.ts files that are not valid TS. For example,
the following:

```
/** @constructor */
function C {}
/** @type {string} */
C.foo;

/** @constructor @extends {C} */
function D {}
/** @type {string|null} */
D.foo;
```
will produce invalid .d.ts because D.foo has to be a subtype of C.foo.

However, the workaround has proved too burdensome and not worth keeping
around. In some cases, we produce up to 3% smaller .d.ts bundles by
removing it.

Also none of this is an issue when using ES6 classes in Closure.