Add support to clutz for generating types for ES modules that use goog.declareModuleId
This makes it possible for a TypeScript file to import an ES Module via its declared module id.

The changes are mostly as you'd expect, with one one hacky part. Polymer currently supports this use case via declaration files that we generate ourselves. We'd want to do this even if clutz was perfect, because we want to use types that are not currently expressible in closure's type system.

However, with this change, clutz would generate types for the Polymer libraries. In codebases include both the clutz types and our custom declarations, the declarations are merged, causing worse types at best, and in some cases new type errors.

My solution here is somewhat hacky: have clutz generate empty declarations for files in third_party/javascript/polymer/v2/...