Add example of union of an interface extending missing base. (#875)
It appears that interface that extends a missing base and is unioned
with null or undefined, then it is collapsed just to a null or
undefined.

Currently, this test is broken, but we know it is a recent change that
broke it. We cannot fix it in Clutz, so we have to wait for a fix in
Closure, in the mean time I am checking in the test so that we can
verify that a fix would works.