Fix imports from files with multiple namespaces. (#884)
Previously, gents assumed that each file only had one namespace that it
exported, so when checking if an import was of a default export or not,
it resolved to whether whatever the first exported namespace in the file
was.  This change makes it so each namespace that's exported is tracked
separately.