Switch to 'export import' for reexporting aliases.
Clutz needs to emit aliases for any sort of symbol, but in TypeScript
aliases can be awkward. If the symbol is a value and a type we have
been using:

type B = ns.A;
val B: typeof ns.A;

This can be often incorrect because in partial mode we don't know if
ns.A is a value and type or just type or just value. However, it appears
specifically for namespaces there is another syntax:

namespace ns2 {
  export import B = ns.A;
}

which is correct for all possible symbols A (even if A is an interface
with generics).

It appears this only works when one accesses through a namespace 'ns',
but in clutz all internal symbols are in the o_O.clutz namespace.

Thanks to @shicks for finding out about 'export import'.