Add a test showing the union typedef collapse feature.
It appears at some point closure has started collapsing typedefs that
only differ by optional fields like:

`A = {t: string, a?: string}` and `B = {t; string, b?: string}`

This results clutz emitting just `A` for the type `A|B`. This is almost
ok, except for the fact that the "unknown object literal" check now
cannot allow assignment of {t: '', b: ''} to what looks like should be
able to accept 'A|B'.

This is adding a test to show this behavior, so that we can notice when
it changes (it apprears to have changed in the last few months).