Removes the type_guard.* test.
Previously this test was using the goog.isDef which was deprecated in
favor of ' !== undefined'. However, since we run this test without
externs, it is currently broken because 'undefined' is *ahem* not
defined to Closure.

Instead, of fixing it yet once again, I noticed that this test doesn't
test anything significant, because the resulting .d.ts doesn't depend on
the function implemenation. So this time I am just deleting it.