Remove destructuring to fix JSC_GOOG_SCOPE_NON_ALIAS_LOCAL
The destructuring pattern this test is using is banned in regular compiler invocations, but until an upcoming Closure Compiler change, is allowed in Clutz invocations.

Also adds a regression test for when the right-hand side of a goog.provide assignment uses a goog.module.get. I ran into this case while testing a Closure Compiler change.