Exporting flat object is not considered default export removal.
Gents needs to remove default exports like `exports = MyCtrl` because they are
not allowed in TypeScript. Moreover, gents communicates that decision to
migrants through log.json.

However, the common pattern of `exports = {A, B, C};` looks like a default
export, but doesn't need default exports removal, as it can simply move the
export to the objects A, B, C, like `export class A`, etc.

I added support for this rewriting
https://github.com/angular/clutz/commit/6a6274481e9241a1ea1eb313aa78073acc4706bd

However, I made a mistake in that I didn't change the log.json emit, so the
rest of the pipeline still thinks that a default export was removed and
replaced with a named export. So to goes on to change the import lines from

const {A, B, C} = goog.require(...);

to

const myModule = goog.require(...);
const {A, B, C} = myModule.newDefaultExportName;

This change adds a test to verify that log.json has an empty defaultName field
(signifying that default exports removal did not happen), and fixes the issue.

PiperOrigin-RevId: 279416612