Change gents to emit interface for @typedef {{a: ..., b:...}} etc.
Turns out migrants was changing these to interfaces already due to autofix in
fix.sh. But the autofix was partial (it did not change the member separators
from , to ; as appropriate for interfaces).

Even that error was fixable by rerunning fix.sh (as we have another linter
autofix for it), but noone known to rerun fix.sh.

Instead of running fix.sh multiple times, we just change gents to emit the
stylistically preferable pattern upfront.

PiperOrigin-RevId: 283871864