Properly emit renaming named imports from TS.
Previously, we would translate incorrectly:

const {a: b} = goog.require('tsprefix.a.b.c');

to

import {a} from 'a/b/c';

Now this is fixed to be

import {a as b} from 'a/b/c';

The fix was already there but there was a needless duplication between the code
that handled imports from already migrated TS, and imports from code that is
co-migrating in the same migrants run.

The common code is now extracted.

PiperOrigin-RevId: 298418241