Fix a crash in function argument comments linking.
The crash happened due to assumption that the nextNode is on the same line as
the comment and substring was called with indices not guaranteed to be in range.

Note that comment attaching is still very error prone, as seen by the test
which adds newlines for unknown reason. At least the crash is taken care of.

PiperOrigin-RevId: 299234939