Gents golden test input files should be a goog.module or goog.provide to match the expected input to gents.  Furthermore, making a test input file a goog.module prevents variable name clashes with other files.
This change causes a golden test to fail if it is not a goog.module or goog.provide because most often this happens when the developer just forgot to mark a test input file as a goog.module.

In the cases where it is explictly necessary to not be a goog.module or goog.provide, the assertion is skipped in the test generators.

PiperOrigin-RevId: 302093871