Add tests for static props on ES6 classes in gents.
These are props declared like this:
```
class C {...}
C.prop = foo;
```

It appears that gents does the wrong thing if there is no JsDoc type. I.e. it
leaves them without further changes which is a TS complier error.

If there is a type depending on the complexity of the initializer it might
either move them to 'static' (preferrable) or declare the prop and leave the
initalizer as is (not looking good, but no type error).

PiperOrigin-RevId: 303210164