All of the Gents files were in a single package.  Because of this, public and package-private members act very similar from a access perspective.  It also makes it harder to follow the code.
With this change, any compiler pass is in the 'pass' package, any
utility class is in the 'util' package, and the rest of the classes
are in the root package.

In the future, these packages can be further subdivided if needed.
That is, if there is a collection of utility classes related to the
same topic, they could occur in a subpackage of 'util' or if there
are several inner classes in a compiler pass class, each of those
inner classes could be in a separate subpackage of the 'pass'
package with package-private access to make each file smaller.

PiperOrigin-RevId: 305505842