The new ExperimentTracker class is used in gents to determine whether or not to use particular experiments.
This will be used in future changes to have an experiment with using
comments attached to nodes by JSCompiler when emitting comments.

Further, now there are two sets of golden tests.  There are
golden tests under the `head` directory that are from Gents
running without experiments enabled and golden tests under the
`experimental` directory that are from Gents running with
experiments enabled.

The `singleTests` and `multiTests` directories were moved under the new `head` directory.

Also, the new `experimental` directory contains copies of these directories.  This is because, in this change, Gents does not have any experiments so the output with and without experiments enabled is the same.

PiperOrigin-RevId: 306310057