The `GENTS_CLANG_FORMAT` environment variable can be used to specify a custom location for clang-format.
The previous code used the `gents.clangFormat` system property.
However, properties are not as nice as environment variables since they
need to be provided to the JVM when starting the app
```
   java -Dmy.property=my.value ...
```
Environment variables, on the other hand can be specified in shell
configuration files or when running an app
```
   MY_PROPERTY=my.value ./some-app
```
That is why this change uses an environment variable instead of a
system property.

PiperOrigin-RevId: 306492474