Prior to this change, the golden tests would contain //!! comments that were used to add comments that described the test.
The testing infrastructure, however, would read a golden file from disk and remove the //!! comments before comparing the golden file contents with the output of gents.

This caused a mismatch when golden tests are regenerated because the regenerated tests (correctly) did not have the //!! comments.

However, a diff of the change made it look like the regeneration of the golden files were incorrect because they lost //!! comments.

To make things easier, the //!! have been replaced with regular // comments since those comments have not special treatment by the testing infrastructure.

PiperOrigin-RevId: 306516222