Prior to this change, the GentsCodeGenerator would give preference to comments identified by the AstCommentLinkingPass compared to the CommentLinkingPass on a per node basis.
With this change, the GentsCodeGenerator gives preference to using a comment
identified by the AstCommentLinkingPass if that pass has associated the comment
with any node in the AST.

This is needed to pave the way for better support for trailing comments.  This
is because the CommentLinkingPass can incorrectly assign a trailing comment to
a node X that occurs earlier in the AST from the node Y it should be associated with
(and which the AstCommentLinkingPass has associated it with).

If precedence of comments is done on a per-node basis, then the comment will be
generted by node X by the GentsCodeGenerator before it could be correctly printed by
Y.

With this change, the comment will not be associated with node X because the
GentsCodeGenerator knows it was associated with another node (Y) by the
AstCommentLinkingPass.  As such, the comment will be associated with node Y which
can process it as a trailing comment after X has been processed.

PiperOrigin-RevId: 310619389