remove --partialInput flag logic
A previous change already made it an error to not provide
the --partialInput flag (to flush out any leftover users):
http://google3/third_party/java_src/clutz/src/main/java/com/google/javascript/clutz/DeclarationGenerator.java?l=355&rcl=308666557

So we now know the flag is always true.
This change deletes the logic that was conditional based
on the flag, and tries to remove references to a world
where we didn't have partial input.

This still leaves removing the flag completely for a different
CL.  We need to adjust callers to not pass the flag, before we
can remove the support for the flag here.  So this also removes
the requirement to pass the flag so we can do that.

PiperOrigin-RevId: 310722217