Emit generic interface and implemented interfaces in generic classes
Previously, gents would not emit interfaces for a generic class, for which gents
provides a custom code generation.

While fixing this, I realized that interface declarations did not have generics
emitted. This is because during conversion to an interface node, the original
`@interface`-annotated JS node's JSDocs are dropped, and the type annotation
pass does not look for generics in interfaces.

By propagating JSDocs for converted interfaces to the annotation pass, picking
up generics for interfaces, and providing a custom codegen emitter for
interfaces with generics this bug is also resolved.

PiperOrigin-RevId: 313445881