Flatten if statement in else body to be in "else if" form
If the body of a conditional branch is not wrapped in an explicit block, Rhino
will add an implicit block around it. For example

```
if (0) {
} else if (1) {
}
```

is represented as

```
if (0) {
} else {*
  if (1) {
  }
*}
```

by the AST, where `{* ... *}` is a synthetic block added to the AST.

In such cases, it make sense for gents to unwrap the syntehtic block because the
author of the original code intended for the body to not be wrapped in a block,
and nested if statements are generally more visually appealing in an "else if"
form.

PiperOrigin-RevId: 314603392