Change gents to emit foo?: string for foo: string | undefined.
Despite looking different, this is actually a better semantic perserving
transformation. Because in closure the following is allowed:

/** @type {{a: string|undefined}} */
const x = {};

and in order to perserve that in TS we have to emit:

const x: {a?: string} = {};

The simpler translation `const x: {a: string|undefined} = {}` is a type error
in TypeScript.

This change handles typedefs, interfaces, records and class members.

Closes #413

PiperOrigin-RevId: 314606112