Support generics on functions and class methods
Before this commit gents only supported emitting generics on class and interface
declarations. Turns out JSComp will handle emitting generics for functions
(which includes free functions and class methods) if function generics are a
Node placed on the `Node.GENERIC_TYPE_LIST` property of the function name. By
modifying the type annotation pass to do this for function-like nodes, we get
the emit for free.

The only special case we need to handle here is if the function is also
abstract, in which case we take over the emit and check if the function name has
generics.

This only handles functions with explicit `@template` annotations, functions that
have implicit templates by virtue of their parameters being templated will be
handled in a separate CL.

PiperOrigin-RevId: 314791683