Add support for emitting an inner function that is nested inside a goog.provided fn.
For example, previously when clutz saw

```
goog.provide('a.b');
a.b = function() {};

a.b.f = function() {}
```
the inner f was not emitted.

We have an inner walk pass, but it was not visiting functions.  Upon adding it,
got the emit, but then we double emitted static functions, and namespace-let
conflict. Two other conditions had to be adjusted to make sure those two cases
are handled.

PiperOrigin-RevId: 314802127