Remove custom code emission for classes and interfaces
JSComp handles emitting class/interface type parameters, "extends" clauses, and
"implements" clauses, so long as class/interface AST nodes are configured in a
certain way. In particular,

- The type parameters are a `GENERIC_TYPE_LIST` node on the `GENERIC_TYPE_LIST`
  property of the class/interface name.
- If a super class in an "extends" clause has type parameters, then the super
  class should be a `PARAMETERIZED_TYPE` node where the first child is the super
  class name and all subsequent children are the type parameters. All children
  of the `PARAMETERIZED_TYPE` node are `NAMED_TYPE`s, which just wrap around a
  name.
  - This does not apply to interface "extends" clauses, which are a bit
    different -- they are stored in an `INTERFACE_EXTENDS` node and already
    properly generated during the type conversion pass.

The only special case that needs to be handled for classes/interfaces is that of
an abstract class, for which TS expects an "abstract" keyword but JSComp
doesn't. In this case gents just checks if "abstract" needs to be emitted and
then hands control over to JSComp for the rest of the class codegen.

PiperOrigin-RevId: 314956182